+++ data/examples/declaration/value/function/arrow/proc-do-complex-out.hs
@@ -9,7 +9,7 @@ foo
       ( (a, b),
         (c, d),
         (e, f)
-        )
+      )
     -> do
       -- Begin do
       (x, y) <- -- GHC parser fails if layed out over multiple lines